requirements.txt
azure-common==1.1.4
azure-nspkg==1.0.0
azure-storage==0.32.0
colorama==0.3.7
Django==1.10.1
futures==3.0.5
numpy==1.11.0
pandas==0.18.1
pdfminer==20140328
Pillow==3.2.0
pkg-resources==0.0.0
PyPDF2==1.26.0
PySimpleSOAP==1.10
python-dateutil==2.5.3
pytz==2016.4
qrcode==5.3
reportlab==3.3.0
requests==2.10.0
six==1.10.0
soap2py==1.16
suds==0.4
xlrd==1.0.0
xmltodict==0.10.1
